# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "HOTEL",
        "description": "API for hotel management",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/rooms": {
            "get": {
                "operationId": "swagTest",
                "tags": [
                    "Rooms"
                ],
                "description": "Recupere la liste des chambres",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "room_number": {
                                        "type": "integer"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "room_type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "operationId": "addRoom",
                "tags": [
                    "Rooms"
                ],
                "description": "Ajoute une chambre",
                "parameters": [
                    {
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "room_number": {
                                    "type": "integer"
                                },
                                "capacity": {
                                    "type": "integer"
                                },
                                "price_per_night": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "status": {
                                    "type": "boolean"
                                },
                                "room_type": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/rooms/{id}": {
            "get": {
                "operationId": "rooms_id",
                "tags": [
                    "Rooms"
                ],
                "description": "Recupere une chambre par son id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_room": {
                                    "type": "integer"
                                },
                                "room_number": {
                                    "type": "integer"
                                },
                                "capacity": {
                                    "type": "integer"
                                },
                                "price_per_night": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "status": {
                                    "type": "boolean"
                                },
                                "room_type": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "patch": {
                "operationId": "updateRoom",
                "tags": [
                    "Rooms"
                ],
                "description": "Met a jour une chambre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "room_number": {
                                    "type": "integer"
                                },
                                "capacity": {
                                    "type": "integer"
                                },
                                "price_per_night": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "status": {
                                    "type": "boolean"
                                },
                                "room_type": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "operationId": "deleteRoom",
                "tags": [
                    "Rooms"
                ],
                "description": "Supprime une chambre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/rooms/pricesAbove/{price}": {
            "get": {
                "operationId": "pricesAbove",
                "tags": [
                    "Rooms"
                ],
                "description": "Recupere les chambres avec un prix superieur a un prix donne",
                "parameters": [
                    {
                        "name": "price",
                        "in": "path",
                        "required": true,
                        "type": "number",
                        "format": "float"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "room_number": {
                                        "type": "integer"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "room_type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/rooms/type/{type}": {
            "get": {
                "operationId": "roomsByType",
                "tags": [
                    "Rooms"
                ],
                "description": "Recupere les chambres d'un type donne",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "room_number": {
                                        "type": "integer"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "room_type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/rooms/pricesBelow/{price}": {
            "get": {
                "operationId": "PricesBelow",
                "tags": [
                    "Rooms"
                ],
                "description": "Recupere les chambres avec un prix inferieur a un prix donne",
                "parameters": [
                    {
                        "name": "price",
                        "in": "path",
                        "required": true,
                        "type": "number",
                        "format": "float"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "room_number": {
                                        "type": "integer"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "room_type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/rooms/prices/{min}/{max}": {
            "get": {
                "operationId": "PricesBetween",
                "tags": [
                    "Rooms"
                ],
                "description": "Recupere les chambres avec un prix compris entre deux prix donnes",
                "parameters": [
                    {
                        "name": "min",
                        "in": "path",
                        "required": true,
                        "type": "number",
                        "format": "float"
                    },
                    {
                        "name": "max",
                        "in": "path",
                        "required": true,
                        "type": "number",
                        "format": "float"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "room_number": {
                                        "type": "integer"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "room_type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}